EXP-445 fix(api): close the pre-release audit findings (#72)
* EXP-444 docs(changelog): write the 4.0.0 section

3.0.0 made the SDK agree with the published contract. This release makes it safe
to operate.

The remaining defects came from a consumer-perspective audit, and not one was a
contract divergence - the document says nothing about any of them. They are about
what happens when a call does not complete, when the provider ships a value this
version has not seen, and when the natural thing to write in C# is the wrong thing.

Held unsigned and untagged until the pre-release audit reports. The section is
written; the tag waits.


* EXP-445 fix(api): close the pre-release audit findings

The release gate said DO NOT SHIP and it was right. Six findings verified and
fixed; the worst was a regression I introduced two releases ago.

EXP-429 taught Product to write quantity, net_amount and vat_amount as JSON
strings, because that is what createPayment declares and the gateway had rejected
numbers there. AlternativePaymentProduct derives from Product, so it inherited
those converters - and the document declares the SAME FIELD NAMES with DIFFERENT
types for that operation: quantity integer, price and the amounts numbers.

Proven live by the auditor: identical body with "quantity":"1" answers
400 invalid_request_body param: products.quantity, with 1 it passes. Every
alternative payment carrying products was broken, by my fix, and my own gate could
not see it because it compares field names and never types across operations.

AlternativePaymentProduct is now declared from the document instead of derived. One
shared base cannot serve two contracts that disagree.

Both refund operations declare PayPartsProduct for their product list, not the
payments-shaped Product; retyped. Batch customer.payment_method declares the same
discriminated union createPayment uses - the legacy flat model sent
{"cc_token":"tok"} where the provider expects {"cc_token":{"token":"tok"}}, so
every tokenized batch payment was off-contract. Gift subscription now uses the
customer type the document requires, which is the only one that can carry the
required payment_method.

EventKey included the operation but not its state, so a pending or 3-D Secure
delivery and the final one for the same operation shared a key. A consumer
following the property's own advice would have rejected the final delivery and
never marked the payment paid - the dedup meant to protect the booking losing it
instead. Status and is_success are now part of the key, with a regression test.

FinMon's recipient IPN was an int. A Ukrainian RNOKPP encodes days since 31.12.1899
in its first five digits, so for anyone born after October 1958 the ten-digit value
exceeds int.MaxValue: the operation was unusable for most living recipients. It is
a long, as the document's unbounded integer allows.

Serialized proof that the two product contracts now diverge correctly:

  alt-payment : {"name":"Consultation","quantity":1,"price":1.00,"net_amount":0.83}
  payments    : {"name":"Consultation","quantity":"1","net_amount":"1.00"}

Suite: 1548 passed on net10.0, 0 failed.


---------

+++ src/SYT.RozetkaPay/Services/FinMonService.cs
@@ -35,7 +35,7 @@ public class FinMonService : BaseService, IFinMonService
     /// <param name="recipientIpn">IPN of recipient</param>
     /// <param name="cancellationToken">Cancellation token</param>
     /// <returns>P2P limits response</returns>
-    public async Task<FinMonP2PPaymentPreLimitsResponse> GetRulesAsync(int recipientIpn, CancellationToken cancellationToken = default)
+    public async Task<FinMonP2PPaymentPreLimitsResponse> GetRulesAsync(long recipientIpn, CancellationToken cancellationToken = default)
     {
         string ipn = recipientIpn.ToString(CultureInfo.InvariantCulture);
         return await GetAsync<FinMonP2PPaymentPreLimitsResponse>(
